backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.54.3
python-docx==0.8.11
PyPDF2==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
